Intrusively discover public API endpoints for a parent domain name
A tool to help you intrusively discover all the publicly available endpoints for all the subdomains of a parent domain name. We first use subdomain enumeration and then combine it with ffuf to fuzz each subdomain for an endpoint with the help of a  custom wordlist to help bruteforce a substantial amount of probable scenarios. Can be modified for the better in the future so stay tuned!